day24

+++ day24/src/bit_array.rs
@@ -1,7 +1,7 @@
 use std::default;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Default)]
-pub struct BitArray(pub u8);
+pub struct BitArray(pub u128);
 
 impl BitArray {
     pub fn new() -> Self {
